# frontend: Improve Map performance