--- a/frontend/src/components/resourceMap/graph/graphFiltering.test.ts
+++ b/frontend/src/components/resourceMap/graph/graphFiltering.test.ts
@@ -40,38 +40,12 @@ describe('filterGraph', () => {
     { id: 'e2', source: '3', target: '4', type: 'kubeRelation' },
   ];
 
-  it('filters nodes by name', () => {
-    const filters: GraphFilter[] = [{ type: 'name', query: 'node1' }];
-    const { nodes: filteredNodes } = filterGraph(nodes, edges, filters);
-
-    // Output contains node1 and node related to it node 2
-    expect(filteredNodes.map(it => it.id)).toEqual(['1', '2']);
-  });
-
   it('filters nodes by namespace', () => {
     const filters: GraphFilter[] = [{ type: 'namespace', namespaces: new Set(['ns3']) }];
     const { nodes: filteredNodes } = filterGraph(nodes, edges, filters);
 
     // Output contains two nodes that both have same namespace ns3
     expect(filteredNodes.map(it => it.id)).toEqual(['3', '4']);
-  });
-
-  it('filters nodes by related nodes', () => {
-    const filters: GraphFilter[] = [{ type: 'related', id: '1' }];
-    const { nodes: filteredNodes } = filterGraph(nodes, edges, filters);
-
-    // Output contains node with id 1 and node 2 that is related to it
-    expect(filteredNodes.map(it => it.id)).toEqual(['1', '2']);
-  });
-
-  it('filters nodes by custom filter function', () => {
-    const filters: GraphFilter[] = [
-      { type: 'custom', label: 'custom', filterFn: node => node.id === '1' },
-    ];
-    const { nodes: filteredNodes } = filterGraph(nodes, edges, filters);
-
-    // Custom filter includes node with id 1 and node 2 that is related to it
-    expect(filteredNodes.map(it => it.id)).toEqual(['1', '2']);
   });
 
   it('filters nodes by error status', () => {