--- a/frontend/src/components/resourceMap/graph/graphFiltering.ts
+++ b/frontend/src/components/resourceMap/graph/graphFiltering.ts
@@ -3,24 +3,11 @@ import { GraphEdge, GraphNode } from './graphModel';
 
 export type GraphFilter =
   | {
-      type: 'name';
-      query: string;
-    }
-  | {
       type: 'hasErrors';
     }
   | {
       type: 'namespace';
       namespaces: Set<string>;
-    }
-  | {
-      type: 'related';
-      id: string;
-    }
-  | {
-      type: 'custom';
-      label: string;
-      filterFn: (node: GraphNode) => boolean;
     };
 
 /**
@@ -31,11 +18,8 @@ export type GraphFilter =
  * even if they don't match the filter
  *
  * The filters can be of the following types:
- * - `name`: Filters nodes by the name
- * - `related`: Keeps only the node with the id and nodes connected to it with edges
  * - `hasErrors`: Filters nodes that have errors based on their resource status. See {@link getStatus}
  * - `namespace`: Filters nodes by their namespace
- * - `custom`: Filters nodes using a custom filter function provided in the filter
  *
  * @param nodes - List of all the nodes in the graph
  * @param edges - List of all the edges in the graph
@@ -80,13 +64,6 @@ export function filterGraph(nodes: GraphNode[], edges: GraphEdge[], filters: Gra
     let keep = true;
 
     filters.forEach(filter => {
-      if (filter.type === 'name' && filter.query.trim().length > 0) {
-        keep &&=
-          'resource' in node.data && node.data.resource?.metadata?.name?.includes(filter.query);
-      }
-      if (filter.type === 'related') {
-        keep &&= node.id === filter.id;
-      }
       if (filter.type === 'hasErrors') {
         keep &&= 'resource' in node.data && getStatus(node?.data?.resource) !== 'success';
       }
@@ -95,9 +72,6 @@ export function filterGraph(nodes: GraphNode[], edges: GraphEdge[], filters: Gra
           'resource' in node.data &&
           !!node.data?.resource?.metadata?.namespace &&
           filter.namespaces.has(node.data?.resource?.metadata?.namespace);
-      }
-      if (filter.type === 'custom') {
-        keep &&= filter.filterFn(node);
       }
     });