--- a/frontend/src/components/resourceMap/graph/graphFiltering.ts
+++ b/frontend/src/components/resourceMap/graph/graphFiltering.ts
@@ -1,4 +1,5 @@
 import { getStatus } from '../nodes/KubeObjectStatus';
+import { makeGraphLookup } from './graphLookup';
 import { GraphEdge, GraphNode } from './graphModel';
 
 export type GraphFilter =
@@ -26,11 +27,17 @@ export type GraphFilter =
  * @param filters - List of fitlers to apply
  */
 export function filterGraph(nodes: GraphNode[], edges: GraphEdge[], filters: GraphFilter[]) {
+  if (filters.length === 0) {
+    return { nodes, edges };
+  }
+
   const filteredNodes: GraphNode[] = [];
   const filteredEdges: GraphEdge[] = [];
 
   const visitedNodes = new Set();
   const visitedEdges = new Set();
+
+  const graphLookup = makeGraphLookup(nodes, edges);
 
   /**
    * Add all the nodes that are related to the given node
@@ -41,21 +48,26 @@ export function filterGraph(nodes: GraphNode[], edges: GraphEdge[], filters: Gra
     if (visitedNodes.has(node.id)) return;
     visitedNodes.add(node.id);
     filteredNodes.push(node);
-    edges.forEach(edge => {
-      if (edge.source === node.id) {
-        const targetNode = nodes.find(it => it.id === edge.target);
-        if (targetNode && !visitedNodes.has(targetNode.id)) pushRelatedNodes(targetNode);
-      }
-      if (edge.target === node.id) {
-        const sourceNode = nodes.find(it => it.id === edge.source);
-        if (sourceNode && !visitedNodes.has(sourceNode.id)) pushRelatedNodes(sourceNode);
-      }
 
-      if (edge.target === node.id || edge.source === node.id) {
+    graphLookup.getOutgoingEdges(node.id)?.forEach(edge => {
+      const targetNode = graphLookup.getNode(edge.target);
+      if (targetNode && !visitedNodes.has(targetNode.id)) {
         if (!visitedEdges.has(edge.id)) {
           visitedEdges.add(edge.id);
           filteredEdges.push(edge);
         }
+        pushRelatedNodes(targetNode);
+      }
+    });
+
+    graphLookup.getIncomingEdges(node.id)?.forEach(edge => {
+      const sourceNode = graphLookup.getNode(edge.source);
+      if (sourceNode && !visitedNodes.has(sourceNode.id)) {
+        if (!visitedEdges.has(edge.id)) {
+          visitedEdges.add(edge.id);
+          filteredEdges.push(edge);
+        }
+        pushRelatedNodes(sourceNode);
       }
     });
   }