This fixes a stack overflow crash #2534 and improves grouping and filtering performance 

There's still a bunch of stuff that can be optimized so this is only first step

Testing done:

- [x] Tested on KWOK cluster with 20,000 pods

After fixes:
Map loading time: 5s
Filter by error: <1s
Group by node: <1s

Before:
Map loading time: ~3.5min
Filter by error: 50s
Group by node: 9s