@illume 2024-11-07
Very nice.

Did you look into how it reacts with a fairly high frequency of change? I mean with pods/events being added, rather than with a static set of resources.
(If not, probably something to look at in another PR.)