@sniok 2024-11-08
> Doh. I think I broke the test plugins CI job with the release of the new headlamp-k8s/eslint-config. It is trying to upgrade to the "latest" `headlamp-k8s/eslint-config`, where it should be getting the version related to that version of headlamp-plugin.
> 
> So you can ignore that test plugins job failure in this PR.

Got it, thanks for the explanation